+++ carp_mobile_sensing/lib/domain/device_deployment.dart
@@ -7,10 +7,10 @@
 
 part of domain;
 
-/// Contains the entire description and configuration for how a CAMS master
-/// device participates in running a study.
+/// Contains the entire description and configuration for how a smartphone master
+/// device participates in deployment of a study.
 @JsonSerializable(fieldRename: FieldRename.none, includeIfNull: false)
-class CAMSMasterDeviceDeployment extends MasterDeviceDeployment {
+class SmartphoneDeployment extends MasterDeviceDeployment {
   late String _studyDeploymentId;
 
   /// The unique id of this study deployment.
@@ -29,16 +29,16 @@ class CAMSMasterDeviceDeployment extends MasterDeviceDeployment {
   /// [DataPointHeader] as the [userId].
   String? userId;
 
-  /// The textual [StudyProtocolDescription] containing the title, description
-  /// and purpose of this study protocol.
-  StudyProtocolDescription? protocolDescription;
+  /// The [StudyDescription] containing the title, description,
+  /// purpose, and the responsible researcher for this study.
+  StudyDescription? protocolDescription;
 
   /// The PI responsible for this study.
-  StudyProtocolReponsible? get responsible => protocolDescription?.responsible;
+  StudyReponsible? get responsible => protocolDescription?.responsible;
 
   SamplingSchemaType? samplingStrategy;
 
-  CAMSMasterDeviceDeployment({
+  SmartphoneDeployment({
     required String studyDeploymentId,
     this.protocolDescription,
     required MasterDeviceDescriptor deviceDescriptor,
@@ -63,9 +63,9 @@ class CAMSMasterDeviceDeployment extends MasterDeviceDeployment {
     _studyDeploymentId = studyDeploymentId;
   }
 
-  CAMSMasterDeviceDeployment.fromMasterDeviceDeployment({
+  SmartphoneDeployment.fromMasterDeviceDeployment({
     required String studyDeploymentId,
-    StudyProtocolDescription? protocolDescription,
+    StudyDescription? protocolDescription,
     required MasterDeviceDeployment masterDeviceDeployment,
   }) : this(
           studyDeploymentId: studyDeploymentId,
@@ -81,12 +81,12 @@ class CAMSMasterDeviceDeployment extends MasterDeviceDeployment {
           dataEndPoint: masterDeviceDeployment.dataEndPoint,
         );
 
-  /// Create a [CAMSMasterDeviceDeployment] based on a [CAMSStudyProtocol].
+  /// Create a [SmartphoneDeployment] based on a [CAMSStudyProtocol].
   /// This method basically makes a 1:1 mapping between a protocol and
   /// a deployment.
-  CAMSMasterDeviceDeployment.fromStudyProtocol({
+  SmartphoneDeployment.fromStudyProtocol({
     required String studyDeploymentId,
-    StudyProtocolDescription? protocolDescription,
+    StudyDescription? protocolDescription,
     required String masterDeviceRoleName,
     DataEndPoint? dataEndPoint,
     required StudyProtocol protocol,
@@ -116,9 +116,9 @@ class CAMSMasterDeviceDeployment extends MasterDeviceDeployment {
     schema.adapt(this, restore: restore);
   }
 
-  factory CAMSMasterDeviceDeployment.fromJson(Map<String, dynamic> json) =>
-      _$CAMSMasterDeviceDeploymentFromJson(json);
-  Map<String, dynamic> toJson() => _$CAMSMasterDeviceDeploymentToJson(this);
+  factory SmartphoneDeployment.fromJson(Map<String, dynamic> json) =>
+      _$SmartphoneDeploymentFromJson(json);
+  Map<String, dynamic> toJson() => _$SmartphoneDeploymentToJson(this);
 
   String toString() => '$runtimeType - studyDeploymentId: $studyDeploymentId, '
       'device: ${deviceDescriptor.roleName}, '