+++ carp_mobile_sensing/lib/domain/domain.dart
@@ -14,7 +14,7 @@ import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 
 // part 'study_protocol.dart';
-part 'study_protocol_description.dart';
+part 'study_description.dart';
 part 'data_endpoint.dart';
 part 'sampling_schema.dart';
 part 'measures.dart';