+++ carp_mobile_sensing/lib/runtime/app_task_executor.dart
@@ -164,6 +164,9 @@ enum UserTaskState {
   /// Done by the user.
   done,
 
+  /// Expired.
+  expired,
+
   /// An undefined state and cannot be used.
   /// Task should be ignored.
   undefined,